chore: reorganize image and video manipulation functions and types
