+++ gallery/src/modules/thumbnails/types/index.ts
@@ -5,11 +5,3 @@ export interface ThumbnailOptions {
   /** Whether to process galleries in subdirectories recursively */
   recursive: boolean;
 }
-
-/** Represents width and height dimensions */
-export interface Dimensions {
-  /** Width in pixels */
-  width: number;
-  /** Height in pixels */
-  height: number;
-}
